Split API webpack into prod and dev variants
